dashboard/requirements.txt
pylint
python-dotenv
streamlit
pandas
altair
boto3
psycopg[binary]
plotly